Move operations to different files
